fix: testìtest:토리 정ë리 (#59)
